fix(sdk-dotnet): Fix join task worker group during boostrap server restart
