+++ sdk-dotnet/LittleHorse.Sdk/Worker/Internal/LHServerConnectionManager.cs
@@ -1,21 +1,14 @@
-﻿using System.Reflection;
-using Google.Protobuf.Collections;
-using Google.Protobuf.WellKnownTypes;
+﻿using Google.Protobuf.Collections;
 using LittleHorse.Sdk.Common.Proto;
-using LittleHorse.Sdk.Exceptions;
-using LittleHorse.Sdk.Helper;
 using Microsoft.Extensions.Logging;
-using Polly;
 using static LittleHorse.Sdk.Common.Proto.LittleHorse;
-using LHTaskException = LittleHorse.Sdk.Exceptions.LHTaskException;
-using TaskStatus = LittleHorse.Sdk.Common.Proto.TaskStatus;
 
 namespace LittleHorse.Sdk.Worker.Internal
 {
     public class LHServerConnectionManager<T> : IDisposable
     {
         private const int BALANCER_SLEEP_TIME = 5000;
-        private const int MAX_REPORT_RETRIES = 5;
+        private const int GRPC_UNARY_CALL_TIMEOUT_SECONDS = 30;
 
         private readonly LHConfig _config;
         private readonly ILogger? _logger;
@@ -23,25 +16,20 @@ namespace LittleHorse.Sdk.Worker.Internal
         private bool _running;
         private List<LHServerConnection<T>> _runningConnections;
         private readonly Thread _rebalanceThread;
-        private readonly SemaphoreSlim _semaphore;
         private readonly LHTask<T> _task;
 
         public LHConfig Config => _config;
         public TaskDef TaskDef => _task.TaskDef!;
 
         public LHServerConnectionManager(LHConfig config,
-                                         LHTask<T> task)
+                                         LHTask<T> task, LittleHorseClient bootstrapClient)
         {
             _config = config;
             _logger = LHLoggerFactoryProvider.GetLogger<LHServerConnectionManager<T>>();
             _task = task;
-            _bootstrapClient = config.GetGrpcClientInstance();
-
+            _bootstrapClient = bootstrapClient;
             _running = false;
             _runningConnections = new List<LHServerConnection<T>>();
-
-            _semaphore = new SemaphoreSlim(config.WorkerThreads);
-
             _rebalanceThread = new Thread(RebalanceWork);
         }
 
@@ -54,39 +42,50 @@ namespace LittleHorse.Sdk.Worker.Internal
         public void Dispose()
         {
             _running = false;
+            GC.SuppressFinalize(this);
         }
 
         private void RebalanceWork()
         {
             while (_running)
             {
-                DoHeartBeat();
                 try
                 {
+                    DoHeartBeat();
                     Thread.Sleep(BALANCER_SLEEP_TIME);
                 }
-                catch { }
+                catch (Exception ex)
+                {
+                    _logger!.LogError(ex, $"Something weird happended {ex.Message}");
+                }
             }
         }
 
         private void DoHeartBeat()
         {
+            var request = new RegisterTaskWorkerRequest
+            {
+                TaskDefId = _task.TaskDef!.Id,
+                TaskWorkerId = _config.WorkerId
+            };
+            _logger?.LogError($"The code reach here 2: {request.TaskWorkerId} and {request.TaskDefId.Name}");
+            var asyncCall = _bootstrapClient.RegisterTaskWorkerAsync(request);
+
             try
             {
-                var request = new RegisterTaskWorkerRequest
-                {
-                    TaskDefId = _task.TaskDef!.Id,
-                    TaskWorkerId = _config.WorkerId,
-                };
-
-                var response = _bootstrapClient.RegisterTaskWorker(request);
-
+                var response = asyncCall.ResponseAsync.Result;
                 HandleRegisterTaskWorkerResponse(response);
             }
             catch (Exception ex)
             {
-                _logger?.LogError(ex, $"Failed contacting bootstrap host {_config.BootstrapHost}:{_config.BootstrapPort}");
-                _runningConnections = new List<LHServerConnection<T>>();
+                _logger?.LogError(ex,
+                    $"Failed contacting bootstrap host {_config.BootstrapHost}:{_config.BootstrapPort}");
+                CloseAllConnections();
+                Thread.Sleep(BALANCER_SLEEP_TIME);
+            }
+            finally
+            {
+                asyncCall.Dispose();
             }
         }
 
@@ -98,8 +97,8 @@ namespace LittleHorse.Sdk.Worker.Internal
                 {
                     try
                     {
-                        var newConnection = new LHServerConnection<T>(this, host);
-                        newConnection.Open();
+                        var newConnection = new LHServerConnection<T>(this, host, _task);
+                        newConnection.Start();
                         _runningConnections.Add(newConnection);
                         _logger?.LogInformation($"Adding connection to: {host.Host}:{host.Port} for task '{_task.TaskDef!.Id}'");
                     }
@@ -124,6 +123,7 @@ namespace LittleHorse.Sdk.Worker.Internal
                     _runningConnections.RemoveAt(i);
                 }
             }
+            _logger!.LogError($"Number of running connections are {_runningConnections.Count}");
         }
 
         private bool ShouldBeRunning(LHServerConnection<T> runningThread, RepeatedField<LHHostInfo> hosts)
@@ -135,158 +135,15 @@ namespace LittleHorse.Sdk.Worker.Internal
         {
             return _runningConnections.Any(conn => conn.IsSame(host.Host, host.Port));
         }
-
-        public async void SubmitTaskForExecution(ScheduledTask scheduledTask)
+        
+        private void CloseAllConnections()
         {
-            await _semaphore.WaitAsync();
-
-            DoTask(scheduledTask);
-        }
-
-        private void DoTask(ScheduledTask scheduledTask)
-        {
-            ReportTaskRun result = ExecuteTask(scheduledTask, LHMappingHelper.MapDateTimeFromProtoTimeStamp(scheduledTask.CreatedAt));
-
-            var wfRunId = LHHelper.GetWfRunId(scheduledTask.Source);
-
-            try
+            _runningConnections.RemoveAll(serverConnection =>
             {
-                var retriesLeft = MAX_REPORT_RETRIES;
-
-                _logger?.LogDebug($"Going to report task for wfRun {wfRunId?.Id}");
-                Policy.Handle<Exception>().WaitAndRetry(MAX_REPORT_RETRIES,
-                    retryAttempt => TimeSpan.FromSeconds(5),
-                    onRetry: (exception, timeSpan, retryCount, context) =>
-                    {
-                        --retriesLeft;
-                        _logger?.LogDebug(
-                            $"Failed to report task for wfRun {wfRunId}: {exception.Message}. Retries left: {retriesLeft}");
-                        _logger?.LogDebug(
-                            $"Retrying reportTask rpc on taskRun {LHHelper.TaskRunIdToString(result.TaskRunId)}");
-                    }).Execute(() => RunReportTask(result));
-            }
-            catch (Exception ex)
-            {
-                _logger?.LogDebug($"Failed to report task for wfRun {wfRunId}: {ex.Message}. No retries left.");
-            }
-            finally
-            {
-                _semaphore.Release();
-            }
-        }
-
-        private void RunReportTask(ReportTaskRun reportedTask)
-        {
-            _bootstrapClient.ReportTask(reportedTask);
-        }
-
-        private ReportTaskRun ExecuteTask(ScheduledTask scheduledTask, DateTime? scheduleTime)
-        {
-            var taskResult = new ReportTaskRun
-            {
-                TaskRunId = scheduledTask.TaskRunId,
-                AttemptNumber = scheduledTask.AttemptNumber
-            };
-
-            var workerContext = new LHWorkerContext(scheduledTask, scheduleTime);
-
-            try
-            {
-                var result = Invoke(scheduledTask, workerContext);
-                var serialized = LHMappingHelper.MapObjectToVariableValue(result);
-
-                taskResult.Output = serialized;
-                taskResult.Status = TaskStatus.TaskSuccess;
-
-                if (!string.IsNullOrEmpty(workerContext.LogOutput))
-                {
-                    var outputVariable = new VariableValue
-                    {
-                        Str = workerContext.LogOutput
-                    };
-
-                    taskResult.LogOutput = outputVariable;
-                }
-
-            }
-            catch (LHInputVarSubstitutionException ex)
-            {
-                _logger?.LogError(ex, "Failed calculating task input variables");
-                taskResult.LogOutput = LHMappingHelper.MapExceptionToVariableValue(ex, workerContext);
-                taskResult.Status = TaskStatus.TaskInputVarSubError;
-                taskResult.Error = new LHTaskError
-                {
-                    Message = ex.ToString(), Type = LHMappingHelper.GetFailureCodeFor(taskResult.Status)
-                };
-            }
-            catch (LHSerdeException ex)
-            {
-                _logger?.LogError(ex, "Failed serializing Task Output");
-                taskResult.LogOutput = LHMappingHelper.MapExceptionToVariableValue(ex, workerContext);
-                taskResult.Status = TaskStatus.TaskOutputSerializingError;
-                taskResult.Error = new LHTaskError
-                {
-                    Message = ex.ToString(), Type = LHMappingHelper.GetFailureCodeFor(taskResult.Status)
-                };
-            }
-            catch (TargetInvocationException ex)
-            {
-                if (ex.GetBaseException() is LHTaskException taskException)
-                {
-                    _logger?.LogError(ex, "Task Method threw a Business Exception");
-                    taskResult.LogOutput = LHMappingHelper.MapExceptionToVariableValue(ex, workerContext);
-                    taskResult.Status = TaskStatus.TaskException;
-                    taskResult.Exception = new Common.Proto.LHTaskException
-                    {
-                        Name = taskException.Name, 
-                        Message = taskException.Message, 
-                        Content = taskException.Content
-                    };
-                }
-                else
-                {
-                    _logger?.LogError(ex, "Task Method threw an exception");
-                    taskResult.LogOutput = LHMappingHelper.MapExceptionToVariableValue(ex, workerContext);
-                    taskResult.Status = TaskStatus.TaskFailed;
-                    taskResult.Error = new LHTaskError
-                    {
-                        Message = ex.InnerException!.ToString(), 
-                        Type = LHMappingHelper.GetFailureCodeFor(taskResult.Status)
-                    };
-                }
-            }
-            catch (Exception ex)
-            {
-                _logger?.LogError(ex, "Unexpected exception during task execution");
-                taskResult.LogOutput = LHMappingHelper.MapExceptionToVariableValue(ex, workerContext);
-                taskResult.Status = TaskStatus.TaskFailed;
-                taskResult.Error = new LHTaskError
-                {
-                    Message = ex.ToString(), Type = LHMappingHelper.GetFailureCodeFor(taskResult.Status)
-                };
-            }
-
-            taskResult.Time = Timestamp.FromDateTime(DateTime.UtcNow);
-
-            return taskResult;
-        }
-
-        private object? Invoke(ScheduledTask scheduledTask, LHWorkerContext workerContext)
-        {
-            var inputs = _task.TaskMethodMappings.Select(mapping => mapping.Assign(scheduledTask, workerContext)).ToArray();
-
-            return _task.TaskMethod!.Invoke(_task.Executable, inputs);
-        }
-
-        public void CloseConnection(string host, int port)
-        {
-            var currConn = _runningConnections.FirstOrDefault(c => 
-                c.IsSame(host, port));
-
-            if (currConn != null)
-            {
-                _runningConnections.Remove(currConn);
-            }
+                serverConnection.Dispose();
+            
+                return true;
+            });
         }
     }
 }
